+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
+v20150402 dev
+-------------
+ - Slug Feature Added
+ - Working Page JavaScript Post Function Completed
+
 v20150401 dev
 -------------
  - Management Enhance
- - Changed Template system to Mako
+ - Changed Template System to Mako
  - Completed Apache License Comments
  - Bugs Fix
  - Font Family Update
