+++ foundation/pyotp.py
@@ -14,7 +14,10 @@
 #   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
-
+#
+#   Edited from pyotp
+#   https://github.com/marcobiscaro2112/pyotp/tree/python3
+#   With Python3 Support
 
 try:
     from urllib.parse import quote
