+++ spirit/nutrition/blossom/framework.htm
@@ -32,7 +32,7 @@
 </head>
 <body>
     ${self.body()}
-    <script src="${management_url('jquery.min.js')}" charset="utf-8"></script><%block name="js"/>
+    <script src="${static_url('jquery.min.js')}" charset="utf-8"></script><%block name="js"/>
 </body>
 </html>
 <%def name="header()">
